requirements.txt
Flask==3.0.0
python-telegram-bot==20.7
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
